backend/requirements.txt
# fastapi==0.110.0
# uvicorn==0.27.1
# python-dotenv==1.0.1
# requests==2.31.0
# chromadb==0.4.24
# langchain==0.1.12
# langchain-community==0.0.28
# langchain-openai==0.1.1
# ollama==0.1.7
# pydantic==2.6.4
# python-multipart==0.0.9

fastapi
uvicorn
python-dotenv
requests
chromadb
langchain
langchain-community
langchain-openai
ollama
pydantic
python-multipart
